Imageprint/requirements.txt
flask 
google-generativeai>=0.3.0 
Pillow
gunicorn